Rename one of the fixture
