+++ application/common/utilities.py
@@ -32,32 +32,3 @@ def get_sequence_number(record: SQSRecord) -> Union[int, None]:
     """
     seq_num_str = record.message_attributes.get("sequence-number", {}).get("stringValue")
     return None if seq_num_str is None else int(seq_num_str)
-
-
-def handle_sqs_msg_attributes(msg_attributes: Dict[str, Any]) -> Dict[str, Any]:
-    attributes = {
-        "error_msg": "",
-        "error_msg_http_code": None,
-        "error_code": "",
-        "rule_arn": "",
-        "target_arn": "",
-        "change_payload": "",
-    }
-    if msg_attributes is not None:
-        if "ERROR_MESSAGE" in msg_attributes:
-            error_msg = msg_attributes["ERROR_MESSAGE"]["stringValue"]
-            attributes["error_msg"] = error_msg
-            error_msg_http_codes = [int(str) for str in error_msg.split() if str.isdigit()]
-            if len(error_msg_http_codes) > 0:
-                attributes["error_msg_http_code"] = error_msg_http_codes[0]
-
-        if "ERROR_CODE" in msg_attributes:
-            attributes["error_code"] = msg_attributes["ERROR_CODE"]["stringValue"]
-
-        if "RULE_ARN" in msg_attributes:
-            attributes["rule_arn"] = msg_attributes["RULE_ARN"]["stringValue"]
-
-        if "TARGET_ARN" in msg_attributes:
-            attributes["target_arn"] = msg_attributes["TARGET_ARN"]["stringValue"]
-
-        return attributes
