+++ application/eventbridge_dlq_handler/eventbridge_dlq_handler.py
@@ -2,7 +2,8 @@ from aws_lambda_powertools import Logger, Tracer
 from aws_lambda_powertools.utilities.data_classes import SQSEvent, event_source
 from aws_lambda_powertools.utilities.typing.lambda_context import LambdaContext
 from common.middlewares import unhandled_exception_logging
-from common.utilities import extract_body, handle_sqs_msg_attributes
+from common.utilities import extract_body
+from typing import Any, Dict
 
 TTL = 157680000  # int((365*5)*24*60*60) . 5 years in seconds
 tracer = Tracer()
@@ -41,3 +42,32 @@ def lambda_handler(event: SQSEvent, context: LambdaContext) -> None:
             "change_payload": body["change_payload"],
         },
     )
+
+
+def handle_sqs_msg_attributes(msg_attributes: Dict[str, Any]) -> Dict[str, Any]:
+    attributes = {
+        "error_msg": "",
+        "error_msg_http_code": None,
+        "error_code": "",
+        "rule_arn": "",
+        "target_arn": "",
+        "change_payload": "",
+    }
+    if msg_attributes is not None:
+        if "ERROR_MESSAGE" in msg_attributes:
+            error_msg = msg_attributes["ERROR_MESSAGE"]["stringValue"]
+            attributes["error_msg"] = error_msg
+            error_msg_http_codes = [int(str) for str in error_msg.split() if str.isdigit()]
+            if len(error_msg_http_codes) > 0:
+                attributes["error_msg_http_code"] = error_msg_http_codes[0]
+
+        if "ERROR_CODE" in msg_attributes:
+            attributes["error_code"] = msg_attributes["ERROR_CODE"]["stringValue"]
+
+        if "RULE_ARN" in msg_attributes:
+            attributes["rule_arn"] = msg_attributes["RULE_ARN"]["stringValue"]
+
+        if "TARGET_ARN" in msg_attributes:
+            attributes["target_arn"] = msg_attributes["TARGET_ARN"]["stringValue"]
+
+        return attributes
